requirements/req.txt
aiohttp==3.9.5
django-storages==1.14.2
channels==4.1.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
argon2-cffi==20.1.0
argon2-cffi-bindings==21.2.0
asgiref
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.7
dataclasses-json==0.6.6
Django
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.6
ecdsa==0.19.0
exceptiongroup==1.2.1
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
gunicorn==22.0.0
hiredis==2.3.2
httplib2==0.22.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-google-community==1.0.5
langchain-text-splitters==0.2.1
langchain-groq==0.1.5
langchain-together==0.1.3
langgraph==0.0.64
langsmith==0.1.75
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.3
packaging==23.2
pluggy==1.5.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
PyJWT==2.8.0
pyparsing==3.1.2
pytest==8.2.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
referencing==0.35.1
requests==2.32.1
rpds-py==0.18.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.30
sqlparse
tenacity==8.3.0
tomli==2.0.1
typing-inspect==0.9.0
typing_extensions
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
yarl==1.9.4
sentry-sdk
django-silk==5.1.0
groq==0.9.0
langchain-groq==0.1.5
boto3==1.34.92
botocore==1.34.92
qrcode==7.4.2
phonenumbers==8.13.40
django-two-factor-auth==1.16.0
pycryptodome==3.20.0
mixpanel==4.10.1